Moving identical code to common module.
